+++ agentkit/__init__.py
@@ -1,3 +1,4 @@
 from .actions import Action  # noqa: F401
 from .actions import ActionException  # noqa: F401
 from .actions.factories.function import action  # noqa: F401
+from agentkit.llms.client.agent import Agent  # noqa: F401
