+++ agentkit/actions/factories/repeat.py
@@ -5,35 +5,26 @@ from agentkit.actions.factories.function import action
 from agentkit.utils.pydantic_utils import create_pydantic_model_from_func
 
 
-def repeat(act: Action, name: str = None, description: str = None, reducer=None) -> Action:
-    if reducer is None:
-        reducer = lambda l: "\n".join([str(e) for e in l])
+def default_reducer(results):
+    return "\n".join(str(e) for e in results)
 
-    def func(*args, **kwargs):
-        if args or len(kwargs) > 1:
+
+def validate_input(args, kwargs, act_name):
+    if args or len(kwargs) > 1:
+        raise ValueError("""
+            Invalid input: The method should not have positional arguments and should only accept one keyword argument""")
+    if kwargs:
+        key, value = next(iter(kwargs.items()))
+        if key != act_name or not isinstance(value, list):
             raise ValueError(
-                f"Invalid input: The method should not have positional arguments and should only accept one keyword argument",
-                f"args: {args}",
-                f"kwargs: {kwargs}",
+                f"Invalid input: The method should accept a single keyword argument '{act_name}' with a list value"
             )
 
-        if kwargs:
-            (key,) = kwargs.keys()
 
-            if key != act.name:
-                raise ValueError(
-                    f"Invalid input: The method should accept a single keyword argument: {act.name}",
-                    f"instead: {key}",
-                )
-
-            (value,) = kwargs.values()
-            if not isinstance(value, list):
-                raise ValueError(
-                    f"Invalid input: The {act.name} should have a list of {act.pydantic_model.__name__} objects",
-                    f"instead: {value}",
-                )
-
-            return reducer([act(**e) for e in value])
+def repeat(act: Action, name: str = None, description: str = None, reducer=default_reducer):
+    def func(*args, **kwargs):
+        validate_input(args, kwargs, act.name)
+        return reducer([act(**e) for e in next(iter(kwargs.values()))] if kwargs else [])
 
     if name is None:
         name = act.name
